pyproject.toml
[tool.poetry]
name = "vtorprice"
version = "0.1.0"
description = ""
authors = ["sergive <vozmiisdelaj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.3"
python-dotenv = "^0.21.0"
drf-yasg = "^1.21.4"
djangorestframework-camel-case = "^1.3.0"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
psycopg2-binary = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"
gunicorn = "^20.1.0"
django-phonenumber-field = "^7.0.0"
phonenumbers = "^8.13.0"
django-mptt = "^0.14.0"
django-mptt-admin = "^2.3.0"
django-admin-multiple-choice-list-filter = "^0.1.1"
django-admin-autocomplete-filter = "^0.7.1"
Pillow = "^9.3.0"
django-model-utils = "^4.3.1"
dadata = "^21.10.1"
pydantic = "^1.10.2"
channels = "^4.0.0"
daphne = "^4.0.0"
channels-redis = "^4.0.0"
django-bulk-update-or-create = "^0.3.0"
drf-nested-routers = "^0.93.4"
shapely = "^2.0.1"
geopy = "^2.3.0"
django-import-export = "^3.2.0"
django-colorfield = "^0.9.0"
python-docx = "^0.8.11"
num2words = "^0.5.12"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
black = "^22.10.0"
autoflake = "^1.7.7"
isort = "^5.10.1"
django-debug-toolbar = "^3.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
